LayoutTests/imported/w3c/web-platform-tests/dom/events/event-global-set-before-handleEvent-lookup.window-expected.txt


PASS window.event is set before 'handleEvent' lookup
